rename: 이미지 디렉토리 변경 (#14)
