+++ src/main/java/com/danthis/backend/domain/classreview/classreviewimage/ClassReviewImage.java
@@ -1,7 +1,7 @@
-package com.danthis.backend.domain.danceclassimage;
+package com.danthis.backend.domain.classreview.classreviewimage;
 
 import com.danthis.backend.domain.BaseEntity;
-import com.danthis.backend.domain.danceclass.DanceClass;
+import com.danthis.backend.domain.classreview.ClassReview;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -16,19 +16,19 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-@Getter
-public class DanceClassImage extends BaseEntity {
+public class ClassReviewImage extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "class_id", nullable = false)
-  private DanceClass danceClass;
+  @JoinColumn(name = "review_id", nullable = false)
+  private ClassReview classReview;
 
   @Column(nullable = false)
   private String imageUrl;
