+++ src/main/java/com/danthis/backend/domain/classreview/classreviewimage/repository/ClassReviewImageRepositoryCustom.java
@@ -1,0 +1,5 @@
+package com.danthis.backend.domain.classreview.classreviewimage.repository;
+
+public interface ClassReviewImageRepositoryCustom {
+
+}
