--- src/main/java/com/danthis/backend/domain/classreviewimage/repository/ClassReviewImageRepositoryCustom.java
@@ -1,5 +1,0 @@
-package com.danthis.backend.domain.classreviewimage.repository;
-
-public interface ClassReviewImageRepositoryCustom {
-
-}
