+++ src/main/java/com/danthis/backend/domain/danceclass/danceclassimage/DanceClassImage.java
@@ -1,7 +1,7 @@
-package com.danthis.backend.domain.classreviewimage;
+package com.danthis.backend.domain.danceclass.danceclassimage;
 
 import com.danthis.backend.domain.BaseEntity;
-import com.danthis.backend.domain.classreview.ClassReview;
+import com.danthis.backend.domain.danceclass.DanceClass;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -16,19 +16,19 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ClassReviewImage extends BaseEntity {
+@Getter
+public class DanceClassImage extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "review_id", nullable = false)
-  private ClassReview classReview;
+  @JoinColumn(name = "class_id", nullable = false)
+  private DanceClass danceClass;
 
   @Column(nullable = false)
   private String imageUrl;
