+++ src/main/java/com/danthis/backend/domain/danceclass/danceclassimage/repository/DanceClassImageRepository.java
@@ -1,7 +1,7 @@
-package com.danthis.backend.domain.danceclassimage.repository;
+package com.danthis.backend.domain.danceclass.danceclassimage.repository;
 
+import com.danthis.backend.domain.danceclass.danceclassimage.DanceClassImage;
 import com.danthis.backend.domain.danceclass.repository.DanceClassRepositoryCustom;
-import com.danthis.backend.domain.danceclassimage.DanceClassImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
