+++ src/main/java/com/danthis/backend/domain/danceclass/danceclassimage/repository/DanceClassImageRepositoryCustom.java
@@ -1,0 +1,5 @@
+package com.danthis.backend.domain.danceclass.danceclassimage.repository;
+
+public interface DanceClassImageRepositoryCustom {
+
+}
