+++ src/main/java/com/danthis/backend/domain/danceclass/danceclassimage/repository/DanceClassImageRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.danthis.backend.domain.danceclassimage.repository;
+package com.danthis.backend.domain.danceclass.danceclassimage.repository;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
