--- src/main/java/com/danthis/backend/domain/danceclassimage/repository/DanceClassImageRepositoryCustom.java
@@ -1,5 +1,0 @@
-package com.danthis.backend.domain.danceclassimage.repository;
-
-public interface DanceClassImageRepositoryCustom {
-
-}
