+++ src/main/java/com/danthis/backend/domain/dancer/Dancer.java
@@ -1,7 +1,7 @@
 package com.danthis.backend.domain.dancer;
 
 import com.danthis.backend.domain.BaseEntity;
-import com.danthis.backend.domain.dancerimage.DancerImage;
+import com.danthis.backend.domain.dancer.dancerimage.DancerImage;
 import com.danthis.backend.domain.mapping.dancergenre.DancerGenre;
 import com.danthis.backend.domain.user.User;
 import jakarta.persistence.*;
@@ -19,36 +19,36 @@ import java.util.Set;
 @Getter
 public class Dancer extends BaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
 
-    @JoinColumn(name="user_id")
-    @OneToOne(fetch = FetchType.LAZY)
-    private User user;
+  @JoinColumn(name = "user_id")
+  @OneToOne(fetch = FetchType.LAZY)
+  private User user;
 
-    @Column(nullable = false, length = 50)
-    private String dancerName;
+  @Column(nullable = false, length = 50)
+  private String dancerName;
 
-    @Column(nullable = false, length = 50)
-    private String instargramId;
+  @Column(nullable = false, length = 50)
+  private String instargramId;
 
-    @Column(nullable = false, length = 10)
-    private String bio;
+  @Column(nullable = false, length = 10)
+  private String bio;
 
-    @Column(nullable = false, length = 10)
-    private String history;
+  @Column(nullable = false, length = 10)
+  private String history;
 
-    private String dancerImage;
+  private String dancerImage;
 
-    private String openChatUrl;
+  private String openChatUrl;
 
-    private Boolean isApproved;
+  private Boolean isApproved;
 
-    @OneToMany(mappedBy = "dancer", fetch = FetchType.LAZY)
-    private Set<DancerGenre> dancerGenres;
+  @OneToMany(mappedBy = "dancer", fetch = FetchType.LAZY)
+  private Set<DancerGenre> dancerGenres;
 
-    @OneToMany(mappedBy = "dancer", fetch = FetchType.LAZY)
-    private Set<DancerImage> dancerImages;
+  @OneToMany(mappedBy = "dancer", fetch = FetchType.LAZY)
+  private Set<DancerImage> dancerImages;
 }
