+++ src/main/java/com/danthis/backend/domain/dancer/dancerimage/DancerImage.java
@@ -1,4 +1,4 @@
-package com.danthis.backend.domain.dancerimage;
+package com.danthis.backend.domain.dancer.dancerimage;
 
 import com.danthis.backend.domain.BaseEntity;
 import com.danthis.backend.domain.dancer.Dancer;
@@ -15,13 +15,13 @@ import lombok.NoArgsConstructor;
 @Getter
 public class DancerImage extends BaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
-    private String imageUrl;
+  private String imageUrl;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "dancer_id")
-    private Dancer dancer;
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "dancer_id")
+  private Dancer dancer;
 }
