+++ src/main/java/com/danthis/backend/domain/dancer/dancerimage/repository/DancerImageRepository.java
@@ -1,9 +1,10 @@
-package com.danthis.backend.domain.dancerimage.repository;
+package com.danthis.backend.domain.dancer.dancerimage.repository;
 
-import com.danthis.backend.domain.dancerimage.DancerImage;
+import com.danthis.backend.domain.dancer.dancerimage.DancerImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface DancerImageRepository extends JpaRepository<DancerImage, Long> {
+
 }
